Save user after modification
More recent versions of Django, such as the one used by Sentry 20,
require an object to be saved before a related object can be updated (if
there are changes pending).

In the case of the LDAP backend the `UserEmail` model is instantiated,
which references the previously unsaved user. Adding `user.save()` after
the `.is_managed` modification solves the issue.